No need for integer conversion

+++ core/src/kvs/surrealcs/cnf.rs
@@ -1,5 +1,4 @@
 use std::sync::LazyLock;
 
-pub static SURREALCS_CONNECTION_POOL_SIZE: LazyLock<i32> =
-	lazy_env_parse_or_else!("SURREAL_SURREALCS_CONNECTION_POOL_SIZE", i32, |_| num_cpus::get()
-		as i32);
+pub static SURREALCS_CONNECTION_POOL_SIZE: LazyLock<usize> =
+	lazy_env_parse_or_else!("SURREAL_SURREALCS_CONNECTION_POOL_SIZE", usize, |_| num_cpus::get());
